backend/requirements.txt
Django==4.1.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.4.0
django-celery-beat==2.8.1
requests==2.31.0
gunicorn==21.2.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
redis==5.0.1
python-dotenv==1.0.0
dj-database-url==2.1.0
cryptography==41.0.7
coverage==7.4.0